requirements.txt
streamlit==1.32.0
pandas==2.1.4
pillow==10.2.0
requests==2.31.0
gspread==5.12.0
oauth2client==4.1.3
